Refactor file and path names for clarity and consistency
