+++ .env.dist
@@ -8,3 +8,5 @@ VITE_SIMPLE_SIGNER_API_FUTURENET=
 # Cypress config
 CYPRESS_HOST_PORT=3001
 CYPRESS_BASE_URL_PREFIX=http://localhost:3000
+VITE_AWS_COGNITO_E2E_EMAIL=test@test.com
+VITE_AWS_COGNITO_E2E_PASSWORD=12345678Test*
