+++ src/config/router/index.tsx
@@ -1,10 +1,6 @@
 import { Navigate, createBrowserRouter } from 'react-router-dom';
 
-import { CollectionVariablesContainer } from '@/common/components/Collections/CollectionVariablesContainer';
-import { AuthProvider } from '@/modules/auth/context/AuthContext';
-import CollectionCTAPage from '@/pages/Collection/CollectionCTAPage';
-import CollectionPage from '@/pages/Collection/CollectionPage';
-import InvocationPage from '@/pages/Invocation/InvocationPage';
+import { CollectionVariablesContainer } from '@/common/components/collections/collectionVariablesContainer';
 import AuthPage from '@/pages/auth/authPage';
 import ChangePassword from '@/pages/auth/changePassword';
 import CreateAccount from '@/pages/auth/createAccount';
@@ -12,14 +8,18 @@ import Login from '@/pages/auth/login';
 import ProtectedRoute from '@/pages/auth/protectedRoute';
 import RecoverPassword from '@/pages/auth/recoverPassword';
 import ResetPassword from '@/pages/auth/resetPassword';
+import CollectionCTAPage from '@/pages/collection/collectionCTAPage';
+import CollectionPage from '@/pages/collection/collectionPage';
+import Home from '@/pages/home/home';
+import InvocationPage from '@/pages/invocation/invocationPage';
+import Providers from '@/providers/providers';
 
 import Root from '@pages/Root';
-import Home from '@pages/home/Home';
 
 const router = createBrowserRouter([
 	{
 		path: '/',
-		element: <AuthProvider />,
+		element: <Providers />,
 		children: [
 			{
 				path: '/',
@@ -87,7 +87,6 @@ const router = createBrowserRouter([
 			},
 		],
 	},
-
 	{
 		path: '*',
 		element: <Navigate to="/" />,
