+++ src/main.tsx
@@ -1,13 +1,9 @@
 import ReactDOM from 'react-dom/client';
 import { RouterProvider } from 'react-router-dom';
 
+import router from './config/router';
 import './index.css';
-import Providers from './providers/Providers';
-
-import router from '@/config/router';
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
-	<Providers>
-		<RouterProvider router={router} />
-	</Providers>,
+	<RouterProvider router={router} />,
 );
