--- src/utils/functions/remove-local-storage-items.ts
@@ -1,7 +1,0 @@
-const removeLocalStorageItems = (...items: string[]) => {
-	items.forEach((item) => {
-		localStorage.removeItem(item);
-	});
-};
-
-export default removeLocalStorageItems;
